+++ src/Heart.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-Heart::Heart(float p_x, float p_y, SDL_Texture* p_tex) : Entity(p_x, p_y, p_tex) {
+Heart::Heart(int p_x, int p_y, SDL_Texture* p_tex) : Entity(p_x, p_y, p_tex) {
     //Initialize the offsets
     xPos = 0;
     yPos = 0;
