+++ src/Tile.cpp
@@ -1,7 +1,7 @@
 #include "Tile.h"
 
 //khởi tạo đối tượng tile có vị trí px,p7 và texture ptex và loại tile ptype
-Tile::Tile(float p_x, float p_y, SDL_Texture* p_tex, int p_type) : Entity(p_x, p_y, p_tex) {
+Tile::Tile(int p_x, int p_y, SDL_Texture* p_tex, int p_type) : Entity(p_x, p_y, p_tex) {
 	//cập nhật collision cho tile
 	collision.x = p_x;
 	collision.y = p_y;
